requirements.txt
# 多模态批处理系统 - 依赖文件
# 云API调用模式

# ============================================
# 核心依赖（必需）
# ============================================
openai>=1.0.0          # OpenAI兼容API调用
pillow>=10.0.0         # 图片处理
colorama>=0.4.6        # 彩色终端输出
pyyaml>=6.0            # 配置文件解析
requests>=2.25.0       # HTTP请求
python-dotenv>=1.0.0   # 环境变量管理（自动加载.env文件）

# ============================================
# Web界面依赖（可选，推荐安装）
# ============================================
gradio>=4.0.0          # Web界面
fastapi>=0.100.0       # API框架
uvicorn>=0.20.0        # ASGI服务器

# ============================================
# 开发工具依赖（可选）
# ============================================
flake8>=6.0.0          # 代码风格检查

# ============================================
# 安装说明
# ============================================
#
# 安装所有依赖：
#   pip install -r requirements.txt
#
# 仅安装核心依赖（不含Web界面）：
#   pip install openai pillow colorama pyyaml requests python-dotenv
